moved ExperimentWizard and some final cleaning
